requirements.txt
PyPDF2>=3.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
anthropic>=0.30.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
